requirements.txt
# Core Dependencies
slack-bolt==1.18.0
openai==1.12.0
langchain==0.1.10
langchain-openai==0.0.5
chromadb==0.4.22
slack-sdk==3.27.1

# AWS SDK
boto3==1.34.34
botocore==1.34.34

# Data Processing
pypdf==4.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# Logging & Monitoring
structlog==24.1.0

# Testing (optional for now)
pytest==8.1.1
pytest-asyncio==0.23.3

# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0

# Vector Database
pinecone-client==3.0.0

